api/requirements.txt
tensorflow-cpu==2.*
keras==2.*
flask==3.*
Flask-Cors==4.*
types-Flask-Cors==4.*
Flask-RESTful==0.*
tinydb==4.*
pydantic==2.*